feat: add Disposable, DisposableSet, AsyncDisposable, AsyncDisposableSet
