+++ src/disposable/createDisposable.ts
@@ -1,0 +1,7 @@
+import { IDisposable } from './types';
+
+export function createDisposable(func: () => void): IDisposable {
+  return {
+    dispose: func,
+  };
+}
