+++ src/disposable/index.ts
@@ -1,3 +1,11 @@
-export * from './disposable';
-export * from './AsyncDisposable';
 export * from './types';
+
+export * from './AsyncDisposable';
+export * from './AsyncDisposableSet';
+
+export * from './Disposable';
+export * from './DisposableSet';
+
+// Older disposable utils
+export * from './createDisposable';
+export * from './DisposableCollection';
